Cleaning, Commenting, and Simplifying towards MVP
Lots of this code is reaching for a bit too much too quickly. Commenting out some bits and stripping away parts we are not ready for yet to just focus on the most basic CRUD tasks of MVP. We can add form validation, authentication, etc. bit by bit over time.